# docs/api/openapi.json
{
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "The request is malformed or fails validation; `fields` names the offending fields, never their values."
      },
      "Conflict" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "A conflict: duplicate, last administrator, or a concurrent change that survived the retries."
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "The caller's role does not allow this (decided by the use case)."
      },
      "NotFound" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Not found - also what an id of another association gets."
      },
      "RuleViolation" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "A business rule refuses it; `message` says which."
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Too many requests; see `Retry-After`."
      },
      "Unauthenticated" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "No valid access token (missing, expired, forged, or the account is no longer active)."
      }
    },
    "schemas" : {
...
      "AcknowledgementResponse" : {
        "properties" : {
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "fields" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
...
      "GroupView" : {
        "properties" : {
          "levels" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "name" : {
            "type" : "string"
          },
          "schedule" : {
            "items" : {
              "$ref" : "#/components/schemas/SlotView"
            },
            "type" : "array"
          },
          "venue" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "JoinAssociationRequest" : {
        "properties" : {
          "consentAccepted" : {
            "type" : "boolean"
          },
          "email" : {
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "phone" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "policyVersion" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "consentAccepted", "email", "name", "phone", "policyVersion" ],
        "type" : "object"
      },
...
      "LevelIdRequest" : {
        "properties" : {
          "levelId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "levelId" ],
        "type" : "object"
      },
      "LevelNameRequest" : {
        "properties" : {
          "name" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "LevelView" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "rank" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "LevelsResponse" : {
        "properties" : {
          "entryLevelId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "levels" : {
            "items" : {
              "$ref" : "#/components/schemas/LevelView"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "PublicAssociationResponse" : {
        "properties" : {
          "contactEmail" : {
            "type" : "string"
          },
          "groups" : {
            "items" : {
              "$ref" : "#/components/schemas/GroupView"
            },
            "type" : "array"
          },
          "levels" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "locality" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "shortName" : {
            "type" : "string"
          },
          "venues" : {
            "items" : {
              "$ref" : "#/components/schemas/VenueView"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "RegisterAssociationRequest" : {
        "properties" : {
          "consentAccepted" : {
            "type" : "boolean"
          },
          "contactEmail" : {
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "founderEmail" : {
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "founderName" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "founderPhone" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "levelNames" : {
            "items" : {
              "maxLength" : 50,
              "minLength" : 0,
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 1,
            "type" : "array"
          },
          "locality" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "nif" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "policyVersion" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "shortName" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "consentAccepted", "contactEmail", "founderEmail", "founderName", "founderPhone", "levelNames", "locality", "name", "policyVersion", "shortName" ],
        "type" : "object"
      },
      "RegisteredAssociationResponse" : {
        "properties" : {
          "shortName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReorderLevelsRequest" : {
        "properties" : {
          "levelIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [ "levelIds" ],
        "type" : "object"
      },
...
      "SlotView" : {
        "properties" : {
          "dayOfWeek" : {
            "type" : "string"
          },
          "durationMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "startTime" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "VenueView" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "courts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The short-lived access token of POST /auth/login or /auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Session-by-session booking for amateur indoor volleyball associations. All paths are under /api/v1. The tenant\n(association) and the caller always come from the bearer token, never from a path, query or body; an id that\nbelongs to another association answers exactly like an id that does not exist (404). Every failure has the body\ndescribed by ApiError. Times are ISO-8601 instants (UTC); weekly schedules and dates are Europe/Lisbon local time.\nMoney is in cents.",
    "title" : "RegiVolley API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/levels" : {
      "post" : {
        "operationId" : "levelAdd",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LevelNameRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LevelsResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "level-controller" ]
      }
    },
    "/api/v1/levels/entry-level" : {
      "put" : {
        "operationId" : "levelEntryLevel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LevelIdRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LevelsResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "level-controller" ]
      }
    },
    "/api/v1/levels/order" : {
      "put" : {
        "operationId" : "levelReorder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReorderLevelsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LevelsResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "level-controller" ]
      }
    },
    "/api/v1/levels/{levelId}" : {
      "put" : {
        "operationId" : "levelRename",
        "parameters" : [ {
          "in" : "path",
          "name" : "levelId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LevelNameRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LevelsResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "level-controller" ]
      }
    },
...
    "/api/v1/public/associations" : {
      "post" : {
        "operationId" : "publicAssociationRegister",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterAssociationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegisteredAssociationResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ ],
        "tags" : [ "public-association-controller" ]
      }
    },
    "/api/v1/public/associations/{shortName}" : {
      "get" : {
        "operationId" : "publicAssociationPage",
        "parameters" : [ {
          "in" : "path",
          "name" : "shortName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicAssociationResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ ],
        "tags" : [ "public-association-controller" ]
      }
    },
    "/api/v1/public/associations/{shortName}/join-requests" : {
      "post" : {
        "operationId" : "publicAssociationJoin",
        "parameters" : [ {
          "in" : "path",
          "name" : "shortName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JoinAssociationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AcknowledgementResponse"
                }
              }
            },
            "description" : "Accepted"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ ],
        "tags" : [ "public-association-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}